pyproject.toml
[project]
name = "neo4j-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.1.0",
    "langfuse>=3.10.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
